Integrated efficient booleans array logic in pucBuffer.
Added functionality for storing the boolean array logic for the
clean/dirty map in a more efficient way by mapping 8 bools on an char
and storing them as char array. The array is stored in the pucBuffer
which is not needed in this function. Added a little service methods for
memory management.