Added temporary SFDP data readout to get supported erase operations.
For smarterase we need to know about the erase operations supported
which are supported by the chip. Because this is not supported by the
spi_flash_types.xml yet there is a function which reads supported erases
from the SFDP and erases flash with them. 
Note: The Erase algorithm is not yet adapted.