[code] added sample config file ( unimplemented )
the sample config file is simmilar to the commandline interface provided by the flasher environment.
It should be possible to freely run tests.
The feature of the black and white mask are powerfull.

bindings for the json are already in the flasher env, have a look art it's argpare feature.