[regression/code] Added global packages
Those python packages have been added in previous tests when
the package.tar.gz was compiled.

those packages are discontinued as a stand alone. this project is the
only one left which makes use of those.
Packages are continued in simpletoolbox.

The test environment can later be added to a butler-test with the
support of simpletoolbox and the flasher handler.