[script] Workaround to allow including base_cli v2.3.
The search for available versions of sub-components on Github is currently based on parsing the release list page.
Only the first page is retrieved and parsed, and older versions that do not appear on the first page are not found.
As a workaround, we add a duplicate of the Github entry with an additional attribute to list only versions older than 2.4.0.