[code][bugfix] When using intflash3, configure intflash1 when needed. (NXTFLASHER-360)
In the commands that configure the flash controller only once according to the start offset, and not for every 4K block (read, verify, is_erased),
make sure that intflash1 is configured when needed.
If Iflash01 (unit 3) is selected and the access starts in intflash0 but ends in intflash1, configure intflash1, too.