[code] Add routines to extract device name and JEDEC ID of an SPI flash from the device description structure. (NXTFLASHER-474)
SpiFlash_getDeviceName() Get the device name from the device description structure.
SpiFlash_getSize()  Get the device size in bytes from the device description structure
SpiFlash_getJedecIdFromIdSeq Try to extract a JEDEC ID from the detection sequence.
We assume that the detection is based on a JEDEC ID if:
- the send/mask/magic bytes for the first byte is 0x9f, 0x00, 0x00
- the send/mask/magic/bytes for each following byte is 0x00, 0xff, 0x??

SpiFlash_getNameAndId() Check the name string. If it starts with SFDP_, assume that the device was detected using the sfdp data. In this case, the name string is based on the JEDEC ID.
If it does not start with SFDP_, we assume that it is a known device detected using the list of known devices. We try to extract a JEDEC ID from this list.