[code] Update behaviour of detect_netx; other commands: exit when not connected (NXTFLASHER-568)
If a netx 90 Rev.1 was detected during Connect(), but no connection has been made (machine interface could not be started), i.e. the exact chip type could not be determined, print the "suspicious" chip type.
detect_netx returns with exit code 2 if the start_MI image has failed during execution, e.g. because the image is not signed, but a signature is required.

All other commands exit if tPlugin:Connect() was not successful.