[code] Preserve the order of helper files. (NXTFLASHER-597)
The list of helper files used to be indexed using the name of the helper file.
This made the entries directly accessible via the name, but had the disadvantage that the order in which the helpers were enumerated during the version check was rather random.
The list is now indexed by integers and the helper name is part of the list entry.
This way, it takes more time to find a helper file, but it should not matter for a small number of helpers.